requirements.txt
flask==3.0.3
pillow==10.3.0
pytesseract==0.3.10
pytest==8.2.0
